macros: Merge ruma-api-macros into ruma-macros
