--- crates/ruma-api-macros/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-categories = ["api-bindings", "web-programming"]
-description = "A procedural macro for generating ruma-api Endpoints."
-homepage = "https://www.ruma.io/"
-keywords = ["matrix", "chat", "messaging", "ruma"]
-license = "MIT"
-name = "ruma-api-macros"
-readme = "README.md"
-repository = "https://github.com/ruma/ruma"
-version = "0.20.0"
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro-crate = "1.0.0"
-proc-macro2 = "1.0.24"
-quote = "1.0.8"
-syn = { version = "1.0.57", features = ["full", "visit"] }
