+++ crates/ruma-macros/src/api/attribute.rs
