+++ crates/ruma-macros/src/api/auth_scheme.rs
