+++ crates/ruma-macros/src/api/response/outgoing.rs
