+++ crates/ruma-macros/src/api/util.rs
