+++ crates/ruma-macros/src/api/version.rs
