+++ crates/ruma-macros/src/lib.rs
@@ -17,6 +17,7 @@ use ruma_identifiers_validation::{
 use syn::{parse_macro_input, DeriveInput, ItemEnum};
 
 use self::{
+    api::{request::expand_derive_request, response::expand_derive_response, Api},
     events::{
         event::expand_event,
         event_content::expand_event_content,
@@ -37,6 +38,7 @@ use self::{
     },
 };
 
+mod api;
 mod events;
 mod identifiers;
 mod serde;
@@ -391,5 +393,39 @@ pub fn derive_string_enum(input: TokenStream) -> TokenStream {
 #[doc(hidden)]
 #[proc_macro_derive(_FakeDeriveSerde, attributes(serde))]
 pub fn fake_derive_serde(_input: TokenStream) -> TokenStream {
+    TokenStream::new()
+}
+
+/// A procedural macro for easily generating [ruma-api]-compatible endpoints.
+///
+/// Note that for technical reasons, the `ruma_api!` macro is only documented in ruma-api, not here.
+///
+/// [ruma-api]: https://github.com/ruma/ruma/tree/main/ruma-api
+#[proc_macro]
+pub fn ruma_api(input: TokenStream) -> TokenStream {
+    let api = parse_macro_input!(input as Api);
+    api.expand_all().into()
+}
+
+/// Internal helper taking care of the request-specific parts of `ruma_api!`.
+#[proc_macro_derive(Request, attributes(ruma_api))]
+pub fn derive_request(input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    expand_derive_request(input).unwrap_or_else(syn::Error::into_compile_error).into()
+}
+
+/// Internal helper taking care of the response-specific parts of `ruma_api!`.
+#[proc_macro_derive(Response, attributes(ruma_api))]
+pub fn derive_response(input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    expand_derive_response(input).unwrap_or_else(syn::Error::into_compile_error).into()
+}
+
+/// A derive macro that generates no code, but registers the ruma_api attribute so both
+/// `#[ruma_api(...)]` and `#[cfg_attr(..., ruma_api(...))]` are accepted on the type, its fields
+/// and (in case the input is an enum) variants fields.
+#[doc(hidden)]
+#[proc_macro_derive(_FakeDeriveRumaApi, attributes(ruma_api))]
+pub fn fake_derive_ruma_api(_input: TokenStream) -> TokenStream {
     TokenStream::new()
 }
